Make it easier to feature-gate entire bogo_shim example
